backend/requirements.txt
scikit-learn
xgboost
numpy
pandas
flask
flask-cors
joblib
catboost
requests
python-dotenv
